requirements.txt
google-cloud-texttospeech
playsound
